dev work on 2/13 from Bruce
