--- HP.Pulsar.KeywordSearch.CommonDataStructures/ProductDataModel.cs
@@ -1,69 +1,0 @@
-﻿namespace HP.Pulsar.KeywordSearch.CommonDataStructures
-{
-    public class ProductDataModel
-    {
-        public int? ProductId { get; set; }
-        public string? ProductName { get; set; }
-        public string? Partner { get; set; }
-        public string? DevCenter { get; set; }
-        public string? Brands { get; set; }
-        public string? SystemBoardId { get; set; }
-        public string? ServiceLifeDate { get; set; }
-        public string? ProductStatus { get; set; }
-        public string? BusinessSegment { get; set; }
-        public string? CreatorName { get; set; }
-        public string? CreatedDate { get; set; }
-        public string? LastUpdaterName { get; set; }
-        public string? LatestUpdateDate { get; set; }
-        public string? SystemManager { get; set; }
-        public string? PlatformDevelopmentPM { get; set; }
-        public string? PlatformDevelopmentPMEmail { get; set; }
-        public string? SupplyChain { get; set; }
-        public string? SupplyChainEmail { get; set; }
-        public string? ODMSystemEngineeringPM { get; set; }
-        public string? ODMSystemEngineeringPMEmail { get; set; }
-        public string? ConfigurationManager { get; set; }
-        public string? ConfigurationManagerEmail { get; set; }
-        public string? CommodityPM { get; set; }
-        public string? CommodityPMEmail { get; set; }
-        public string? Service { get; set; }
-        public string? ServiceEmail { get; set; }
-        public string? ODMHWPM { get; set; }
-        public string? ODMHWPMEmail { get; set; }
-        public string? ProgramOfficeProgramManager { get; set; }
-        public string? ProgramOfficeProgramManagerEmail { get; set; }
-        public string? Quality { get; set; }
-        public string? QualityEmail { get; set; }
-        public string? PlanningPM { get; set; }
-        public string? PlanningPMEmail { get; set; }
-        public string? BIOSPM { get; set; }
-        public string? BIOSPMEmail { get; set; }
-        public string? SystemsEngineeringPM { get; set; }
-        public string? SystemsEngineeringPMEmail { get; set; }
-        public string? MarketingProductMgmt { get; set; }
-        public string? MarketingProductMgmtEmail { get; set; }
-        public string? ProcurementPM { get; set; }
-        public string? ProcurementPMEmail { get; set; }
-        public string? SWMarketing { get; set; }
-        public string? SWMarketingEmail { get; set; }
-        public string? ProductFamily { get; set; }
-        public string? ODM { get; set; }
-        public string? ReleaseTeam { get; set; }
-        public string? RegulatoryModel { get; set; }
-        public string? Releases { get; set; }
-        public string? Description { get; set; }
-        public string? ProductLine { get; set; }
-        public string? PreinstallTeam { get; set; }
-        public string? MachinePNPID { get; set; }
-        public string? EndOfProduction { get; set; }
-        public string? WHQLstatus { get; set; }
-        public string? LeadProduct { get; set; }
-        public string? Chipsets { get; set; }
-        public string? ComponentItems { get; set; }
-        public string? ProductGroups { get; set; }
-        public string? CurrentBIOSVersions { get; set; }
-        public string? Target { get; set; }
-        public int? ID { get; set; }
-
-    }
-}
