--- HP.Pulsar.KeywordSearch.CommonDataStructures/PulsarEnvironment.cs
@@ -1,10 +1,0 @@
-﻿namespace HP.Pulsar.KeywordSearch.CommonDataStructures
-{
-    public enum PulsarEnvironment
-    {
-        Dev,
-        Test,
-        Sandbox,
-        Production
-    }
-}
