--- HP.Pulsar.KeywordSearch.CommonDataStructures/SearchType.cs
@@ -1,13 +1,0 @@
-﻿namespace HP.Pulsar.KeywordSearch.CommonDataStructures
-{
-    public enum SearchType
-    {
-        Product,
-        Root,
-        Version,
-        Feature,
-        PRL,
-        DCR,
-        SuddenImpact
-    }
-}
