--- HP.Pulsar.KeywordSearch.DataReader/HP.Pulsar.KeywordSearch.DataReader.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Data.SqlClient" Version="4.8.5" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.CommonDataStructures\HP.Pulsar.KeywordSearch.CommonDataStructures.csproj" />
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.DataTransformation\HP.Pulsar.KeywordSearch.DataTransformation.csproj" />
-  </ItemGroup>
-
-</Project>
