--- HP.Pulsar.KeywordSearch.Orchestrator/HP.Pulsar.KeywordSearch.Orchestrator.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.CommonDataStructures\HP.Pulsar.KeywordSearch.CommonDataStructures.csproj" />
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.DataReader\HP.Pulsar.KeywordSearch.DataReader.csproj" />
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.DataTransformation\HP.Pulsar.KeywordSearch.DataTransformation.csproj" />
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.DataWriter\HP.Pulsar.KeywordSearch.DataWriter.csproj" />
-  </ItemGroup>
-
-</Project>
