--- HP.Pulsar.KeywordSearch.Orchestrator/ProductOrchestrator.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HP.Pulsar.KeywordSearch.Orchestrator
-{
-    internal class ProductOrchestrator : IInitializationOrchestrator
-    {
-        public ProductOrchestrator()
-        {
-
-        }
-
-        public void Init()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
