--- HP.Pulsar.KeywordSearch.Orchestrator/RootOrchestrator.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HP.Pulsar.KeywordSearch.Orchestrator
-{
-    internal class RootOrchestrator : IInitializationOrchestrator
-    {
-
-        public void Init()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
