--- HP.Pulsar.KeywordSearch.Orchestrator/Search.cs
@@ -1,22 +1,0 @@
-﻿using HP.Pulsar.KeywordSearch.CommonDataStructures;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HP.Pulsar.KeywordSearch.Orchestrator
-{
-    public class Search
-    {
-        public Search()
-        {
-                
-        }
-
-        public IEnumerable<KeywordSearchOutputModel> search(string input)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
