--- HP.Pulsar.KeywordSearch.Orchestrator/Update.cs
@@ -1,17 +1,0 @@
-﻿using HP.Pulsar.KeywordSearch.CommonDataStructures;
-
-namespace HP.Pulsar.KeywordSearch.Orchestrator
-{
-    public class Update
-    {
-        public Update()
-        {
-
-        }
-
-        public void update(SearchType type, int id)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
