+++ HP.Pulsar.Search.Keyword/AppData/App.Beta.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+    <DatabaseConnectionString>server=TdcPulsarItgDb.tpc.rd.hpicorp.net;initial catalog=PRS;integrated security=SSPI</DatabaseConnectionString>
+    <MeilisearchAPIEndpoint></MeilisearchAPIEndpoint>
+</configuration>
