+++ HP.Pulsar.Search.Keyword/CommonDataStructure/SearchType.cs
@@ -1,0 +1,12 @@
+﻿namespace HP.Pulsar.Search.Keyword.CommonDataStructure;
+
+public enum SearchType
+{
+    Product,
+    Root,
+    Version,
+    Feature,
+    PRL,
+    DCR,
+    SuddenImpact
+}
