+++ HP.Pulsar.Search.Keyword/Infrastructure/PulsarEnvironment.cs
@@ -1,0 +1,9 @@
+﻿namespace HP.Pulsar.Search.Keyword.Infrastructure;
+
+public enum PulsarEnvironment
+{
+    Dev,
+    Test,
+    Sandbox,
+    Production
+}
