+++ HP.Pulsar.Search.Keyword/Orchestrator/IInitializationOrchestrator.cs
@@ -1,0 +1,6 @@
+﻿namespace HP.Pulsar.Search.Keyword.Orchestrator;
+
+internal interface IInitializationOrchestrator
+{
+    public Task InitializeAsync();
+}
