+++ HP.Pulsar.Search.Keyword/Orchestrator/ProductOrchestrator.cs
@@ -1,0 +1,21 @@
+﻿namespace HP.Pulsar.Search.Keyword.Orchestrator;
+
+internal class ProductOrchestrator : IInitializationOrchestrator
+{
+    public ProductOrchestrator()
+    {
+
+    }
+
+    public Task InitializeAsync()
+    {
+        // read products from database
+
+        // data processing
+
+        // write to meiliesearch
+
+
+        throw new NotImplementedException();
+    }
+}
