+++ HP.Pulsar.Search.Keyword/Orchestrator/Search.cs
@@ -1,0 +1,16 @@
+﻿using HP.Pulsar.Search.Keyword.CommonDataStructure;
+
+namespace HP.Pulsar.Search.Keyword.Orchestrator;
+
+public class Search
+{
+    public Search()
+    {
+
+    }
+
+    public IEnumerable<KeywordSearchOutputModel> search(string input)
+    {
+        throw new NotImplementedException();
+    }
+}
