+++ HP.Pulsar.Search.Keyword/Orchestrator/Update.cs
@@ -1,0 +1,14 @@
+﻿namespace HP.Pulsar.Search.Keyword.Orchestrator;
+
+public class Update
+{
+    public Update()
+    {
+
+    }
+
+    public void update(SearchType type, int id)
+    {
+        throw new NotImplementedException();
+    }
+}
