+++ KeywordSearch.sln
@@ -3,17 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HP.Pulsar.KeywordSearch.CommonDataStructures", "HP.Pulsar.KeywordSearch.CommonDataStructures\HP.Pulsar.KeywordSearch.CommonDataStructures.csproj", "{3F57077C-9CF3-4496-A0E0-DCFE05C967A0}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HP.Pulsar.KeywordSearch.DataReader", "HP.Pulsar.KeywordSearch.DataReader\HP.Pulsar.KeywordSearch.DataReader.csproj", "{51E12C56-DF36-45D9-9509-C5AB864B1C4F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HP.Pulsar.KeywordSearch.DataTransformation", "HP.Pulsar.KeywordSearch.DataTransformation\HP.Pulsar.KeywordSearch.DataTransformation.csproj", "{DBB954CF-065B-4CEC-8122-2ADD68F02043}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HP.Pulsar.KeywordSearch.DataWriter", "HP.Pulsar.KeywordSearch.DataWriter\HP.Pulsar.KeywordSearch.DataWriter.csproj", "{02B6E23F-4422-4B42-AA6B-1229AA8AC5CD}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KeywordSearchInitializationConsole", "KeywordSearchInitializationConsole\KeywordSearchInitializationConsole.csproj", "{21640116-2ED1-42DD-A9ED-38B65E6D9B8D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HP.Pulsar.KeywordSearch.Orchestrator", "HP.Pulsar.KeywordSearch.Orchestrator\HP.Pulsar.KeywordSearch.Orchestrator.csproj", "{6896844E-E48C-43D5-9219-0A983CC68D96}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HP.Pulsar.Search.Keyword", "HP.Pulsar.Search.Keyword\HP.Pulsar.Search.Keyword.csproj", "{95BD5F71-759C-4EC8-B0D6-CC6FEAF6D7DC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,30 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3F57077C-9CF3-4496-A0E0-DCFE05C967A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F57077C-9CF3-4496-A0E0-DCFE05C967A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F57077C-9CF3-4496-A0E0-DCFE05C967A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F57077C-9CF3-4496-A0E0-DCFE05C967A0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{51E12C56-DF36-45D9-9509-C5AB864B1C4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{51E12C56-DF36-45D9-9509-C5AB864B1C4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{51E12C56-DF36-45D9-9509-C5AB864B1C4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{51E12C56-DF36-45D9-9509-C5AB864B1C4F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DBB954CF-065B-4CEC-8122-2ADD68F02043}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DBB954CF-065B-4CEC-8122-2ADD68F02043}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DBB954CF-065B-4CEC-8122-2ADD68F02043}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DBB954CF-065B-4CEC-8122-2ADD68F02043}.Release|Any CPU.Build.0 = Release|Any CPU
-		{02B6E23F-4422-4B42-AA6B-1229AA8AC5CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{02B6E23F-4422-4B42-AA6B-1229AA8AC5CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{02B6E23F-4422-4B42-AA6B-1229AA8AC5CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{02B6E23F-4422-4B42-AA6B-1229AA8AC5CD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{21640116-2ED1-42DD-A9ED-38B65E6D9B8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{21640116-2ED1-42DD-A9ED-38B65E6D9B8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{21640116-2ED1-42DD-A9ED-38B65E6D9B8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{21640116-2ED1-42DD-A9ED-38B65E6D9B8D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6896844E-E48C-43D5-9219-0A983CC68D96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6896844E-E48C-43D5-9219-0A983CC68D96}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6896844E-E48C-43D5-9219-0A983CC68D96}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6896844E-E48C-43D5-9219-0A983CC68D96}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95BD5F71-759C-4EC8-B0D6-CC6FEAF6D7DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95BD5F71-759C-4EC8-B0D6-CC6FEAF6D7DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95BD5F71-759C-4EC8-B0D6-CC6FEAF6D7DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95BD5F71-759C-4EC8-B0D6-CC6FEAF6D7DC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
