+++ KeywordSearchInitializationConsole/KeywordSearchInitializationConsole.csproj
@@ -1,14 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\HP.Pulsar.KeywordSearch.Orchestrator\HP.Pulsar.KeywordSearch.Orchestrator.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\HP.Pulsar.Search.Keyword\HP.Pulsar.Search.Keyword.csproj" />
+    </ItemGroup>
 
 </Project>
