+++ Source.Properties.target
@@ -1,0 +1,16 @@
+<Project>
+    <PropertyGroup>
+        <Authors>HP Inc.</Authors>
+        <Company>HP Inc.</Company>
+        <Product>HP Pulsar</Product>
+        <LangVersion>10.0</LangVersion>
+        <OutputPath>$(pulsarRoot)\out\$(Configuration)</OutputPath>
+        <Version Condition=" '$(BUILD_BUILDNUMBER)' == '' ">0.0.0.0</Version>
+        <Version Condition=" '$(BUILD_BUILDNUMBER)' != '' ">$(BUILD_BUILDVERSION)</Version>
+    </PropertyGroup>
+    
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        <DefineConstants>DEBUG;TRACE</DefineConstants>
+        <Optimize>false</Optimize>
+    </PropertyGroup>
+</Project>
